+++ packages/core/inline-edit/docs/1-inline-editable-textfield.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import { md, Example, Props, code } from '@atlaskit/docs';
 
 export default md`