+++ packages/core/select/src/components/input-options.js
@@ -1,6 +1,7 @@
 // @flow
-
-import React, { Component, type Element } from 'react';
+/** @jsx jsx */
+import { jsx } from '@emotion/core';
+import { Component, type Element } from 'react';
 import RadioIcon from '@atlaskit/icon/glyph/radio';
 import CheckboxIcon from '@atlaskit/icon/glyph/checkbox';
 import { colors, themed, gridSize } from '@atlaskit/theme';