+++ packages/elements/mention/src/util/i18n.tsx
@@ -30,3 +30,13 @@ export const DefaultAdvisedAction = noPropFormatter(
 );
 export const LoginAgain = noPropFormatter(messages.loginAgain);
 export const DifferentText = noPropFormatter(messages.differentText);
+export const SpotlightTitle = noPropFormatter(messages.spotlightTitle);
+export const SpotlightCloseTooltip = noPropFormatter(
+  messages.spotlightCloseButtonToolTip,
+);
+export const SpotlightDescription = noPropFormatter(
+  messages.spotlightDescription,
+);
+export const SpotlightDescriptionLink = noPropFormatter(
+  messages.spotlightDescriptionLink,
+);