+++ packages/search/global-search/src/components/confluence/ConfluenceFilterGroup.tsx
@@ -6,8 +6,11 @@ import { FormattedMessage } from 'react-intl';
 import { Filter } from '../../api/CrossProductSearchClient';
 import ConfluenceSpaceFilter from './SpaceFilter';
 import styled from 'styled-components';
+import { fireMoreFiltersButtonClickEvent } from '../../util/analytics-event-helper';
+import { CreateAnalyticsEventFn } from '../analytics/types';
+import { withAnalyticsEvents } from '@atlaskit/analytics-next';
 
-interface Props {
+export interface Props {
   spaceAvatar: string;
   spaceTitle: string;
   spaceKey: string;
@@ -15,27 +18,54 @@ interface Props {
   isFilterOn: boolean;
   onFilterChanged(filter: Filter[]): void;
   onAdvancedSearch(event: CancelableEvent): void;
+  createAnalyticsEvent?: CreateAnalyticsEventFn;
+  searchSessionId: string;
 }
 
 const Container = styled.span`
   display: inline-flex;
   align-items: center;
+  max-width: 100%;
 `;
 
-export default class ConfluenceFilterGroup extends React.Component<Props> {
+const ButtonContainer = styled.div`
+  flex-shrink: 0;
+`;
+
+const FilterContainer = styled.div`
+  flex-shrink: 1;
+  overflow: hidden;
+`;
+
+export class ConfluenceFilterGroup extends React.Component<Props> {
+  onMoreFiltersClick = (event: CancelableEvent) => {
+    const {
+      onAdvancedSearch,
+      createAnalyticsEvent,
+      searchSessionId,
+    } = this.props;
+    fireMoreFiltersButtonClickEvent(searchSessionId, createAnalyticsEvent);
+    onAdvancedSearch(event);
+  };
+
   render() {
-    const { onAdvancedSearch } = this.props;
     return (
       <ResultItemGroup
         title={<FormattedMessage {...messages.confluence_space_filter} />}
       >
         <Container>
-          <ConfluenceSpaceFilter {...this.props} />
-          <Button appearance="link" onClick={onAdvancedSearch}>
-            <FormattedMessage {...messages.confluence_more_filters} />
-          </Button>
+          <FilterContainer>
+            <ConfluenceSpaceFilter {...this.props} />
+          </FilterContainer>
+          <ButtonContainer>
+            <Button appearance="link" onClick={this.onMoreFiltersClick}>
+              <FormattedMessage {...messages.confluence_more_filters} />
+            </Button>
+          </ButtonContainer>
         </Container>
       </ResultItemGroup>
     );
   }
 }
+
+export default withAnalyticsEvents<Props>()(ConfluenceFilterGroup);