+++ packages/core/table-tree/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/table-tree
 
+## 7.0.7
+
+- Updated dependencies [790e66bece](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/790e66bece):
+  - @atlaskit/button@13.0.11
+  - @atlaskit/select@10.0.0
+
 ## 7.0.6
 
 - Updated dependencies [06326ef3f7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/06326ef3f7):